PyTorch/requirements.txt
aniso8601==9.0.1
certifi==2021.5.30
charset-normalizer==2.0.2
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
Pillow==8.3.1
pytz==2021.1
requests==2.26.0
six==1.16.0
torch==1.8.0
torchaudio==0.8.0
torchvision==0.9.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
uwsgi==2.0.19.1